# swagger.yaml
swagger: "2.0"
info:
  description: "User Details management API used to perform Retrieve and Update operations corresponding to User Details"
  version: "1.0.0"
  title: "Swagger User Management Microservice"
  contact:
    email: "kpvsekhar@gmail.com"
basePath: "/api/usermgmt/userdetails/"
schemes:
- "http"
securityDefinitions:
  basicAuth:
    type: basic
security:
  - basicAuth: []
paths:
  /retrieve:
    post:
      tags:
      - "retrieveUserDetails"
      summary: "Retrieve User Details"
      operationId: "retrieveUserDetails"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "RetrieveUserDetailsRequest for which user details should be retrieved"
        required: true
        schema:
          $ref: "#/definitions/RetrieveUserDetailsRequest"
      responses:
        "200":
          description: "Successful opertaion"
          schema:
              $ref: "#/definitions/UserDetails"
        "400":
          description: "Bad request - Validation error in the request sent"
          schema:
              $ref: "#/definitions/ValidationErrorResponse"
        "401":
          description: "Authentication and Authorization error"
          schema:
              $ref: "#/definitions/UserManagementErrorRespone"
        "500":
          description: "Internal Server error - Error during the request processing"
          schema:
              $ref: "#/definitions/UserManagementErrorRespone"
  /update:
    post:
      tags:
      - "updateUserDetails"
      summary: "Update User Details"
      operationId: "updateUserDetails"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "userDetails to update the userdetails"
        required: true
        schema:
          $ref: "#/definitions/UserDetails"
      responses:
        "200":
          description: "Successful opertaion"
        "401":
          description: "Authentication and Authorization error"
          schema:
              $ref: "#/definitions/UserManagementErrorRespone"
        "422":
          description: "Unprocessable entity - Given user doesn't exist in the DB to update"
          schema:
              $ref: "#/definitions/UserManagementErrorRespone"
        "500":
          description: "Internal Server error - Error during the request processing"
          schema:
              $ref: "#/definitions/UserManagementErrorRespone"            
definitions:
  RetrieveUserDetailsRequest:
    type: "object"
    properties:
      userId:
        type: "string"
  Address:
    type: "object"
    properties:
      street:
        type: "string"
      city:
        type: "string"
      state:
        type: "string"  
        enum:
        - "NSW"
        - "ACT"
        - "VIC"
        - "QLD"
        - "NT"
        - "SA"
        - "TAS"
        - "WA"
      postcode:
        type: "number"
        format: "int64"
  UserDetails:
    type: "object"
    properties:
      userId:
        type: "number"
        format: "int64"
      title:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      emailId:
        type: "string"
      gender:
        type: "string"
      address:
        $ref: "#/definitions/Address"
  UserManagementErrorRespone:
    type: "object"
    properties:
      message:
        type: "string"
  ValidationErrorResponse:
    type: "object"
    properties:
      errorField:
        type: "string"
      errorMessage:
        type: "string"